Fix reference loop causing memory leak
This fixes a quite substantial memory leak, leaking the entire grayscale
image for every created image.

By taking a reference to newImage in the closure passed to SetFinalizer
a reference loop is created via the finalizer, causing it to never be
freed.

Fix this by using the correct reference passed to the closure by
SetFinalizer.